# Optimize TagBuilder tag generation